Move functions to functions.py
